Re) {style}: Separate Header components
