+++ src/components/Main/Home/HomeHeader.module.scss
@@ -1,0 +1,33 @@
+.locationBox {
+  padding: 2px;
+  display: flex;
+  margin-right: calc(100% - 200px);
+
+  .location {
+    margin: 0;
+    padding-top: 4px;
+    padding-left: 15px;
+    font-size: 20px;
+    font-weight: bold;
+  }
+
+  .locationArrow {
+    position: relative;
+    height: 15px;
+
+    top: 50%;
+    transform: translate(0, -50%);
+
+    padding-left: 5px;
+  }
+}
+
+.headerImg {
+  position: relative;
+  top: 50%;
+  transform: translate(0, -50%);
+
+  height: 25px;
+
+  margin-right: 10px;
+}
