+++ src/components/Main/Main.tsx
@@ -5,37 +5,14 @@ import Chat1 from "../../Image/Footer/chat-selected.png";
 import Chat2 from "../../Image/Footer/chat-unselected.png";
 import User1 from "../../Image/Footer/user-selected.png";
 import User2 from "../../Image/Footer/user-unselected.png";
-import DownArrow from "../../Image/Header/down-arrow.png";
-import Search from "../../Image/Header/search.png";
-import Category from "../../Image/Header/category.png";
-import Notice from "../../Image/Header/bell.png";
-import { useEffect, useState } from "react";
-import MainGoods from "./MainGoods";
+import {useEffect, useState} from "react";
+import HomeGoods from "./Home/HomeGoods";
+import HomeHeader from "./Home/HomeHeader";
 
 const Main = () => {
-  const [location, setLocation] = useState("");
+  const [location, setLocation] = useState("301동");
   const [writeHandle, setWriteHandle] = useState(false);
   const [page, setPage] = useState("home");
-
-  useEffect(() => {
-    setLocation("301동");
-  });
-
-  const handleLocation = () => {
-    console.log("위치 재지정하는 팝업");
-  };
-
-  const handleSearch = () => {
-    console.log("검색창으로 push");
-  };
-
-  const handleCategory = () => {
-    console.log("카테고리 선책창으로 push");
-  };
-
-  const handleNotice = () => {
-    console.log("알림창으로 push");
-  };
 
   const changeToHome = () => {
     setPage("home");
@@ -55,47 +32,20 @@ const Main = () => {
 
   return (
     <>
+      <div className={styles.test}>
+      </div>
       <div
         className={`${styles.backShadow} ${writeHandle ? styles.show : ""}`}
       />
       <div className={styles.wrapper}>
         <div className={styles.header}>
-          {page === "home" && (
-            <>
-              <div className={styles.locationBox} onClick={handleLocation}>
-                <p className={styles.location}>{location}</p>
-                <img
-                  className={styles.locationArrow}
-                  src={DownArrow}
-                  alt="화살표"
-                />
-              </div>
-              <img
-                className={styles.headerImg}
-                src={Search}
-                onClick={handleSearch}
-                alt="검색"
-              />
-              <img
-                className={styles.headerImg}
-                src={Category}
-                onClick={handleCategory}
-                alt="카테고리"
-              />
-              <img
-                className={styles.headerImg}
-                src={Notice}
-                onClick={handleNotice}
-                alt="알림"
-              />
-            </>
-          )}
+          {page === "home" && <HomeHeader location={location} setLocation={setLocation}/>}
           {page === "chat" && <></>}
           {page === "user" && <></>}
         </div>
         <div className={styles.contents}>
           {page === "home" && (
-            <MainGoods
+            <HomeGoods
               writeHandle={writeHandle}
               setWriteHandle={setWriteHandle}
             />
